requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
gunicorn==20.0.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.11.15
prompt-toolkit==3.0.14
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.7.4
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
virtualenv==20.4.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.2.0
